[README]
I manually merged art with design and got everything back in working order. Gerad set up this New-Master branch to act as a replacement of the broken one.

Here is some vital information.
DO NOT MERGE OLD BRANCHES WITH NEW-MASTER.

Only work with the new branches that will be made once this is uploaded.