Created a global shader script
It works technically. The issue with it is that we don't have any appropriate shaders to apply globally that wont horribly break other stuff, such as emissive materials and the water shader.

The alternative is to just do it by hand, which is too much work, and possible performance heavy.

Regardless, I made the script so I could even find out the info above.