Minor Update
Fixed a few issues with the visuals, such as things like being able to out of bounds.

Applied the new environment textures to the levels.

Started fixing the resets in level 3, but that damn piston bug is back at it again.

Didn't get much done yet, but I mean, its something.